# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    AgentRole:
      type: string
      enum: [ROUTA, CRAFTER, GATE, DEVELOPER]
    ModelTier:
      type: string
      enum: [SMART, BALANCED, FAST]
    AgentStatus:
      type: string
      enum: [PENDING, ACTIVE, COMPLETED, ERROR, CANCELLED]
...
    Agent:
      type: object
      required: [id, name, role, modelTier, workspaceId, status, createdAt, updatedAt, metadata]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/AgentRole"
        modelTier:
          $ref: "#/components/schemas/ModelTier"
        workspaceId:
          type: string
        parentId:
          type: string
        status:
          $ref: "#/components/schemas/AgentStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
          const: true
...
paths:
...
  /api/agents:
    get:
      operationId: listAgents
      summary: List agents (or get single by id query param)
      parameters:
        - name: id
          in: query
          schema:
            type: string
          description: If provided, returns single agent
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/AgentRole"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AgentStatus"
        - name: parentId
          in: query
          schema:
            type: string
        - name: summary
          in: query
          schema:
            type: string
          description: If present with id, returns agent summary (Next.js only)
      responses:
        "200":
          description: Agent list or single agent
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      agents:
                        type: array
                        items:
                          $ref: "#/components/schemas/Agent"
                  - $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name:
                  type: string
                role:
                  $ref: "#/components/schemas/AgentRole"
                workspaceId:
                  type: string
                  default: "default"
                parentId:
                  type: string
                modelTier:
                  $ref: "#/components/schemas/ModelTier"
      responses:
        "200":
          description: Created agent
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentId:
                    type: string
                  agent:
                    $ref: "#/components/schemas/Agent"
  /api/agents/{id}:
    get:
      operationId: getAgent
      summary: Get agent by ID (REST-style path param)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAgent
      summary: Delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
  /api/agents/{id}/status:
    post:
      operationId: updateAgentStatus
      summary: Update agent status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/AgentStatus"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
                    const: true

  # ── Tasks ──
...
  /api/skills/clone:
    get:
      operationId: discoverSkills
      summary: Discover skills from repo path
      parameters:
        - name: repoPath
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Discovered skills
    post:
      operationId: cloneSkillRepo
      summary: Clone a skill repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                skillsDir:
                  type: string
      responses:
        "200":
          description: Cloned
...
  /api/clone:
    get:
      operationId: listClonedRepos
      summary: List cloned repositories
      responses:
        "200":
          description: Repo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  repos:
                    type: array
                    items:
                      type: object
    post:
      operationId: cloneRepo
      summary: Clone a GitHub repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
      responses:
        "200":
          description: Cloned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  path:
                    type: string
                  name:
                    type: string
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
                  existed:
                    type: boolean
    patch:
      operationId: switchBranch
      summary: Switch branch on cloned repo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repoPath, branch]
              properties:
                repoPath:
                  type: string
                branch:
                  type: string
      responses:
        "200":
          description: Switched
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
...
  /api/clone/branches:
    get:
      operationId: getBranches
      summary: Get branch info
      parameters:
        - name: repoPath
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Branch info
    post:
      operationId: fetchBranches
      summary: Fetch remote branches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repoPath:
                  type: string
      responses:
        "200":
          description: Fetched
    patch:
      operationId: checkoutBranch
      summary: Checkout branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repoPath:
                  type: string
                branch:
                  type: string
                pull:
                  type: boolean
      responses:
        "200":
          description: Checked out

  # ── A2A ──
...